+++ Otherfiles/version_check.py
@@ -4,7 +4,7 @@ import os
 import sys
 import codecs
 Failed = 0
-VERSION = "2.2"
+VERSION = "2.3"
 
 
 SETUP_ITEMS = [